wip: Update jpeg decoder.
I was trying to see an xray that I got recently (happy to email it
privately if it helps), which fails to decode on the demo with:

```
JPX Error: Unsupported COD options (selectiveArithmeticCodingBypass) jpx.js:339:23
```

So this tries to update to openjpeg via
https://github.com/mozilla/pdf.js/pull/17946, which should support it.
However:

 * It's very WIP (the image decodes correctly on the worker, but I only
   see black on the canvas, however I see correct metadata).
 * I haven't updated all the references to the other decoders.
 * It loads the non-jpeg decoders as a module which I can't test, but
   probably breaks them. I also need to update the sync decoders, which
   likely doesn't work.

I don't think I'll have time to finish this, but posting it here in case
it saves someone time or someone has the time to finish it up.

Thanks.